refactor: extract NextStepsConsole and RuledOutConsole from IncidentCard
Reduces IncidentCard.tsx bloat by moving NextStepsConsole (~200 lines)
and RuledOutConsole (~80 lines) into their own files with all their
helper functions.
